Python/Covid-Tracker/requirements.txt
beautifulsoup4==4.9.3
tk==0.1.0
requests==2.25.0
Pillow==8.2.0